restructure: FolderModal 하위 컴포넌트 폴더 구조 변경
- subscribe section과 member section을 다른 폴더로 분류함
