+++ src/component/content/setting-folder/folder-modal/member-section/MemberBox.tsx
@@ -1,4 +1,4 @@
-import { InvitedMember } from "../../../layout/sidebar/SideBarType";
+import { InvitedMember } from "../../../../layout/sidebar/SideBarType";
 
 type Props = {
   member: InvitedMember;
