+++ src/component/content/setting-folder/folder-modal/member-section/MemberSearchList.tsx
@@ -1,4 +1,4 @@
-import { InvitedMember } from "../../../layout/sidebar/SideBarType";
+import { InvitedMember } from "../../../../layout/sidebar/SideBarType";
 import MemberSearchBox from "./MemberSearchBox";
 
 type Props = {
